refactor: context를 root 내부로 통합시켰습니다.
